requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib_metadata==8.7.0
Jinja2==3.1.3
jinja2_fragments==1.3.0
Mako==1.3.2
mangum==0.17.0
MarkupSafe==2.1.4
packaging==23.2
passlib==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyMySQL==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.7
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
sqlparse==0.5.3
starlette==0.35.1
tabulate==0.9.0
typing_extensions==4.9.0
uvicorn==0.26.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yoyo-migrations==9.0.0
zipp==3.23.0
